feat: move ClusterVariableRecord to its own package
