+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/clustervariable/DbClusterVariableState.java
@@ -19,7 +19,7 @@ import io.camunda.zeebe.db.impl.DbString;
 import io.camunda.zeebe.engine.state.clustervariable.DbClusterVariableScopeKey.Scope;
 import io.camunda.zeebe.engine.state.mutable.MutableClusterVariableState;
 import io.camunda.zeebe.protocol.ZbColumnFamilies;
-import io.camunda.zeebe.protocol.impl.record.value.variable.ClusterVariableRecord;
+import io.camunda.zeebe.protocol.impl.record.value.clustervariable.ClusterVariableRecord;
 import java.util.Optional;
 import org.agrona.DirectBuffer;
